Add procedure for deploying to AUR when not on Archlinux
The current instructions in the README for deploying to AUR only work if
you are on Archlinux yourself. If you are on a different distro like
Ubuntu, the deployment must be done from within an Archlinux based
Docker container. This can be a bit of a hassle to set up, so this
commit adds a procedure for automating most of that.

Specifically this commit adds:
- A Dockerfile that defines an image that can be used for deployment.
- Instructions in README for how to use the image for deployment.
- New CI job for testing that the image can be successfully built.