Update lint & test archlinux image
It seems that archlinux/base got killed, so we must use
archlinux/archlinux:base instead.

The new version of shellcheck seems to want to follow any "source"
command to lint the sources files as well. It then complains on the line
`source $VENV/bin/activate` because this file only exists when actually
running the script. In such cases it is recommended to merely disable
this check on that line (https://www.shellcheck.net/wiki/SC1091), which
we do in this commit.